Artie and the Wolf Moon
There should be a big red flag on audiobooks for graphic novels. I had
no idea when I checked this out from the library for my son and I to
listen to. I was baffled for the first half wondering where all the
details were amid the barking and oof-ing.