Add All Hugo Novellas
Realizing that I’ve read at least one of them (and I think several more).
Having them in the document helps when searching. I’m also open to
trying to read all of them at some point, so fleshing out the list.

Normalizing to Wikipedia standard: using italics when it was published
stand-alone and quotation marks when only included in a larger volume.